Add client-side JS files
So far I've been testing the backend of the app. Right now, the app should work just as the original did since I added the front-end JS files. One thing I did notice is the time it takes to get the json with the data seems significantly shorter than while running on rails.